Add tests for time utils

+++ PeriodicBackupService.Core/src/TimeUtils.cs
@@ -2,17 +2,17 @@
 {
 	public class TimeUtils
 	{
-		public static double HoursToMillis(int hours)
+		public static double HoursToMillis(double hours)
 		{
 			return hours * 60 * 60 * 1000;
 		}
 
-		public static double MinutesToMillis(int minutes)
+		public static double MinutesToMillis(double minutes)
 		{
 			return minutes * 60000;
 		}
 
-		public static int MillisToMinutes(int millis)
+		public static double MillisToMinutes(double millis)
 		{
 			return millis / 60000;
 		}
